Changed the variable type of NIM

+++ src/Tugas2/Nilai.java
@@ -1,6 +1,6 @@
 package Tugas2;
 public class Nilai {
-    int NIM;
+    long NIM;
     String nama;
     float NilaiAbsen;
     float NilaiTugas;
